+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1]
+## [0.2.2]
 ### Added
-- Config file is now now a parameter.
+- Config file is now an optional parameter.
+- Log level choice.
+
+## [0.2.1]
+Test release
 
 ## [0.2.0]
 ### Fixed